backend/requirements.txt
django ~= 4.0
python-dotenv ~= 0.19
requests ~= 2.27
djangorestframework
pillow
django-cors-headers

django ~= 5.0.3
requests ~= 2.31.0
django-environ ~= 0.11.2
authlib ~= 1.3.0
pyjwt[crypto]>=2.8.0
cryptography>=41.0.0